+++ src/main/java/com/likelion/DSFest/entity/Review.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 public class Review extends BaseEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer reviewId; //글 id
+    private Long reviewId; //글 id
 
     private String content; // 글 내용
 }
